alias: don't add --sigint-ignore to mysql
or it won't kill ongoing queries. Need a better way to cope with the
racing condition...

Also clear the "nv" alias, since I'm not going to use it in the
foreseeable future because of its breaking changes (e.g. default
'directory' value has changed to one directory, breaking a lot of use
cases of mine).